rename mod `map` to `mapped`
