+++ bevy_asset_loader/src/lib.rs
@@ -73,8 +73,8 @@ pub mod asset_collection;
 pub mod dynamic_asset;
 /// A game state responsible for loading assets
 pub mod loading_state;
-/// Trait definition for mapped assets
-pub mod map;
+/// Trait definition for mapped assets collection
+pub mod mapped;
 /// Dynamic assets for common Bevy asset types
 #[cfg_attr(docsrs, doc(cfg(feature = "standard_dynamic_assets")))]
 #[cfg(feature = "standard_dynamic_assets")]
@@ -97,7 +97,7 @@ pub mod prelude {
             DynamicAssets,
         },
         loading_state::{LoadingState, LoadingStateAppExt, LoadingStateSet},
-        map::{FileName, FileStem, MapKey},
+        mapped::{FileName, FileStem, MapKey},
     };
 }
