+++ bevy_asset_loader/src/mapped.rs
